rf: some refactoring for internals
